Renamed position pages, shortened position url
