+++ src/components/Card.tsx
@@ -14,7 +14,7 @@ export function Card({card}: ModelCardProps) {
     }
 
     return (
-        <Link className="card-item" target="_blank" to={'/position/' + card.id + '/show'}>
+        <Link className="card-item" target="_blank" to={ '/position/' + card.id }>
             <div>
                 <p className="cart-item__title">{card.type}</p>
                 <p className="card-item__text"><span>Марка:</span> {card.mark}</p>
