--- src/components/CardIdViewPrint.tsx
@@ -1,5 +1,0 @@
-export function CardIdViewPrint() {
-    return (
-        <h1>HELLO PRINT CARD</h1>
-    )
-}
