+++ src/components/PositionNavigation.tsx
@@ -5,9 +5,10 @@ export function PositionNavigation() {
     return (
         <nav>
             <span className='block'>
-                <Link className='sidebar-btn' to='show'>Информация</Link>
-                <Link className='sidebar-btn' to='' >Отгрузка</Link>
+                <Link className='sidebar-btn' to=''>Информация</Link>
+                <Link className='sidebar-btn' to='dispatch' >Отгрузка</Link>
                 <Link className='sidebar-btn' to='print'>Печать</Link>
+                <Link className='sidebar-btn' to=''>История</Link>
             </span>
         </nav>
     )
