+++ src/components/PositionPrint.tsx
@@ -1,0 +1,5 @@
+export function PositionPrint() {
+    return (
+        <h1>HELLO PRINT CARD</h1>
+    )
+}
